+++ src/components/Modal/SignInModal/index.jsx
@@ -8,7 +8,7 @@ import {
   faFacebookSquare,
   faGooglePlusG
 } from "@fortawesome/free-brands-svg-icons";
-import { signIn } from '../../../redux/actions/user'
+import { signIn } from '../../../redux/actions/user.signin'
 
 
 class SignInModal extends Component {
